feat(courier-navigation-bar): migrated the folder and resolved the conflict.
